Refactoring Hyphy classes for less redundancy
A lot of code was repeated in each of the child classes, even though no
changes were made. This causes a lot of problems when debugging one or
more of the modules.
In addition, I added some bugfixes for Modeltest and added a properly
parsed output for Modeltest and REL.