changes to Build.PL; added Build.pm and Build/Test.pm
Bio::Tools::Run::Build subclasses Module::Build to add
an interactive tool install selector and methods to
allow checking/install of prereqs only for selected
tools. Test.pm provides a convenience method for skipping
all tests for a given tool - if the tool is not selected
(information from the feature() method of the build
object), the entire test suite is skipped.