maint: remove mention of non-existent Bio::Run from Build.PL (issue 57)
module_name value for Module::Build is set to Bio::Run but there is
no such module.

module_name is used by Module::Build to discover the dist_name and
dist_version.  However both of those are specified explicitly so
module_name is ignored.  Remove because it's wrong and pointless.